Moving more files into folders
