+++ src/python/New paper figures/log3_log2_frac.py
