+++ src/python/New paper figures/longplot.py
@@ -7,7 +7,7 @@ values=[0 for _ in base]
 
 def read_file():
     # Open the file for reading
-    with open("..\\..\\data\\a034258-lpsolve.txt", "r") as file:
+    with open("..\\..\\..\\data\\a034258-lpsolve.txt", "r") as file:
         # Read the file line by line
         for line in file:
             # Split the line into parts (assuming numbers are separated by whitespace)
