+++ src/python/New paper figures/newplot.py
